requirements.txt
discord.py==1.7.3
pytz==2021.3
requests==2.26.0
